@@ -32,6 +32,8 @@
     "class-transformer": "^0.3.2",
     "class-validator": "^0.13.1",
     "dotenv": "^8.2.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.22.1",
     "halson": "^3.0.0",
     "mysql2": "^2.2.5",
     "npx": "^10.2.2",
@@ -59,7 +61,7 @@
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.6.1",
     "@typescript-eslint/parser": "^4.6.1",
-    "eslint": "^7.12.1",
+    "eslint": "^7.21.0",
     "eslint-config-prettier": "7.0.0",
     "eslint-plugin-prettier": "^3.1.4",
     "jest": "^26.6.3",
