@@ -3,4 +3,4 @@ export const DEVELOPMENT = 'development';
 export const TEST = 'test';
 export const PRODUCTION = 'production';
 export const ADMIN_REPOSITORY = "ADMIN_REPOSITORY";
-export const CONTACT_REPOSITORY = "ADMIN_REPOSITORY";
+export const CONTACT_REPOSITORY = "CONTACT_REPOSITORY";
