@@ -1,6 +1,5 @@
 import { Sequelize } from 'sequelize-typescript';
 import { Admin } from 'src/modules/admins/admin.entity';
-import { AdminsModule } from 'src/modules/admins/admins.module';
 import { Contact } from 'src/modules/contacts/contact.entity';
 
 import { SEQUELIZE, DEVELOPMENT, TEST, PRODUCTION } from '../constants';
@@ -24,6 +23,7 @@ export const databaseProviders = [{
         default:
            config = databaseConfig.development;
         }
+        config['logging'] = true;
         const sequelize = new Sequelize(config);
         sequelize.addModels([Admin, Contact]);
         await sequelize.sync();
