@@ -1,7 +1,7 @@
-import { Module } from "@nestjs/common";
-import { AdminsController } from "./admins.controller";
-import { adminsProviders } from "./admins.providers";
-import { AdminsService } from "./admins.service";
+import {Module} from "@nestjs/common";
+import {AdminsController} from "./admins.controller";
+import {adminsProviders} from "./admins.providers";
+import {AdminsService} from "./admins.service";
 
 @Module({
     controllers: [AdminsController],
