@@ -1,5 +1,5 @@
-import { ADMIN_REPOSITORY } from "src/core/constants";
-import { Admin } from "./admin.entity";
+import {ADMIN_REPOSITORY} from "src/core/constants";
+import {Admin} from "./admin.entity";
 
 export const adminsProviders = [{
     provide: ADMIN_REPOSITORY,
