@@ -1,9 +1,8 @@
-import { Inject, Injectable, Req } from "@nestjs/common";
-import { ADMIN_REPOSITORY } from "src/core/constants";
-import { Admin } from "./admin.entity";
-import { CreateAdminDto } from "./dto/create-admin.dto";
-import { Request, Response } from "express";
-import { UpdateAdminDto } from "./dto/update-admin.dto";
+import {Inject, Injectable} from "@nestjs/common";
+import {ADMIN_REPOSITORY} from "src/core/constants";
+import {Admin} from "./admin.entity";
+import {CreateAdminDto} from "./dto/create-admin.dto";
+import {UpdateAdminDto} from "./dto/update-admin.dto";
 
 @Injectable()
 export class AdminsService {
@@ -13,27 +12,31 @@ export class AdminsService {
 
 
     async create(admin: CreateAdminDto): Promise<Admin> {
-        return await this.adminRepository.create<Admin>(admin);
+        return this.adminRepository.create<Admin>(admin);
     }
 
-    //async findAll(req: Request, res: Response, page?: number, rows?: number): Promise<Admin[]> {
-    async findAll(page?: number, rows?: number): Promise<Admin[]> {
-                return await this.adminRepository.findAll<Admin>();
+    async findAll(page?: string, rows?: string): Promise<Admin[]> {
+                let result = await this.adminRepository.findAndCountAll({limit: +rows, offset: +rows*(+page-1)})
+                return result.rows
     }
 
     async findOne(id: number): Promise<Admin> {
-        return await this.adminRepository.findOne<Admin>({ where: { id }});
+        return this.adminRepository.findOne<Admin>({where: {id}});
     }
 
     async findOneByLogin(login: string): Promise<Admin> {
-        return await this.adminRepository.findOne<Admin>({where: {login}});
+        return this.adminRepository.findOne<Admin>({where: {login}});
     }
 
     async update(id: number, admin: UpdateAdminDto): Promise<[number, Admin[]]> {
-        return await this.adminRepository.update<Admin>(admin, {where: { id }});
+        return this.adminRepository.update<Admin>(admin, {where: {id}});
     }
 
     async remove(id: number) {
-        return await this.adminRepository.destroy({where: { id }});
+        return this.adminRepository.destroy({where: {id}});
+    }
+
+    async searchOne(pattern: {}): Promise<Admin> {
+        return this.adminRepository.findOne(pattern);
     }
 }
