@@ -1,4 +1,4 @@
-import { ApiHideProperty, ApiProperty, ApiPropertyOptional, PartialType } from "@nestjs/swagger";
-import { CreateAdminDto } from "./create-admin.dto";
+import {PartialType} from "@nestjs/swagger";
+import {CreateAdminDto} from "./create-admin.dto";
 
 export class UpdateAdminDto extends PartialType(CreateAdminDto) {}
