@@ -1,5 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
+import { CONTACT_REPOSITORY } from 'src/core/constants';
 import { ContactsController } from './contacts.controller';
+import { contactsProviders } from './contacts.provider';
 import { ContactsService } from './contacts.service';
 
 describe('ContactsController', () => {
@@ -8,7 +10,9 @@ describe('ContactsController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [ContactsController],
-      providers: [ContactsService],
+      providers: [
+        ContactsService,
+      ],
     }).compile();
 
     controller = module.get<ContactsController>(ContactsController);
