@@ -1,11 +1,13 @@
-import { Module } from '@nestjs/common';
-import { ContactsService } from './contacts.service';
-import { ContactsController } from './contacts.controller';
-import { contactsProviders } from './contacts.provider';
+import {Module} from '@nestjs/common';
+import {ContactsService} from './contacts.service';
+import {ContactsController} from './contacts.controller';
+import {contactsProviders} from './contacts.provider';
+import {AdminsService} from '../admins/admins.service';
+import {adminsProviders} from '../admins/admins.providers';
 
 @Module({
   controllers: [ContactsController],
   exports: [ContactsService],
-  providers: [ContactsService, ...contactsProviders]
+  providers: [ContactsService, ...contactsProviders, AdminsService, ...adminsProviders]
 })
 export class ContactsModule {}
